pyproject.toml
[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
asyncio_mode="auto"

[tool.poetry]
name = "trade-fastapi"
version = "0.1.0"
description = ""
authors = ["smirnyagga <nomsgva@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
sqlalchemy = "^2.0.23"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.3.6"
flower = "^2.0.1"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
websockets = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
